Note:This revision was used to build 0.8.5-3 Windows package. Previous test result is cleared in GUI when running new tests now. Windows: Vastly improved installer, supports built-in GTK+ now. Windows: A proper method of privilege escalation is used for Vista/7. Windows: GSmartControl works on Windows 7 now. Updated copyright for files with Whatever License to use Unlicense.

+++ gsmartcontrol/src/hz/format_unit.h
@@ -89,27 +89,33 @@ inline std::string format_size(uint64_t size, bool use_decimal = false, bool siz
 	const int addn = static_cast<int>(!use_decimal) + (static_cast<int>(size_is_bits) * 2);
 
 	if (size >= eb_size) {  // exa
-		return hz::number_to_string(static_cast<long double>(size) / eb_size, 2) + HZ_RC_("filesize", names[(6 * 4) + addn]);
+		return hz::number_to_string(static_cast<long double>(size) / static_cast<long double>(eb_size), 2)
+				+ HZ_RC_("filesize", names[(6 * 4) + addn]);
 	}
 
 	if (size >= pb_size) {  // peta
-		return hz::number_to_string(static_cast<long double>(size) / pb_size, 2) + HZ_RC_("filesize", names[(5 * 4) + addn]);
+		return hz::number_to_string(static_cast<long double>(size) / static_cast<long double>(pb_size), 2)
+				+ HZ_RC_("filesize", names[(5 * 4) + addn]);
 	}
 
 	if (size >= tb_size) {  // tera
-		return hz::number_to_string(static_cast<long double>(size) / tb_size, 2) + HZ_RC_("filesize", names[(4 * 4) + addn]);
+		return hz::number_to_string(static_cast<long double>(size) / static_cast<long double>(tb_size), 2)
+				+ HZ_RC_("filesize", names[(4 * 4) + addn]);
 	}
 
 	if (size >= gb_size) {  // giga
-		return hz::number_to_string(static_cast<long double>(size) / gb_size, 2) + HZ_RC_("filesize", names[(3 * 4) + addn]);
+		return hz::number_to_string(static_cast<long double>(size) / static_cast<long double>(gb_size), 2)
+				+ HZ_RC_("filesize", names[(3 * 4) + addn]);
 	}
 
 	if (size >= mb_size) {  // mega
-		return hz::number_to_string(static_cast<long double>(size) / mb_size, 2) + HZ_RC_("filesize", names[(2 * 4) + addn]);
+		return hz::number_to_string(static_cast<long double>(size) / static_cast<long double>(mb_size), 2)
+				+ HZ_RC_("filesize", names[(2 * 4) + addn]);
 	}
 
 	if (size >= kb_size) {  // kilo
-		return hz::number_to_string(static_cast<long double>(size) / kb_size, 2) + HZ_RC_("filesize", names[(1 * 4) + addn]);
+		return hz::number_to_string(static_cast<long double>(size) / static_cast<long double>(kb_size), 2)
+				+ HZ_RC_("filesize", names[(1 * 4) + addn]);
 	}
 
 	return hz::number_to_string(size) + HZ_RC_("filesize", names[(0 * 4) + addn]);
